Refactor great circles/delocalized using new code
This becomes trivial if I promote the Gaussian Oscillator in the
DatasetGenerator to a public property, so that inheriting classes can
simply override it.

OscillatingGeneratorBase is now completely obsolete, since Great Circles
and Delocalized Oscillators no longer inherit from a common base class.
(This is just as well, since my old implementation was almost certainly
wrong.)